pyproject.toml
[project]
name = "thumbnail-alchemist"
version = "0.1.0"
description = "Multimodal AI agent that turns raw video, images, and creator profiles into high-impact YouTube thumbnails"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "pydantic>=2.0.0",
    "uvicorn[standard]",
    "google-cloud-storage>=3.7.0",
    "google-genai>=0.4.0",  # Official Gemini API (replaces deprecated google-generativeai)
    "openai>=1.0.0",
    "librosa>=0.10.2,<0.11.0",
    "numpy",
    "opencv-python>=4.8.0",
    "mediapipe>=0.10.0",
    "python-multipart>=0.0.6",
    "onnxruntime>=1.16.0",
    "huggingface-hub>=0.19.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "torchcodec>=0.1.0",  # For speech detection with Silero VAD
    "soundfile>=0.12.0",
    "pillow>=11.3.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "ruff",
    "pytest",
    "httpx",
    "anyio",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
